Make a common Generator.Frontend dll for CLI tool and BuildTool
Update READMEs
